Add indirections for CoinRules contract key (#39)
This is in preparation for introducing interfaces for CoinRules &
friends. Given that we don’t have interface keys, we have to have a
plain template with the key that can then point to the underlying
interface cid.

I’m not touching mining rounds & validator rewards for now. The reason
for that is that these are not as much exposed to the users (they’re
primarily used from within the internals of CoinRules) so I think we
can make progress on interfacifying a bunch of stuff without having to
touch this.