Drop interface key workaround for CoinRules (#56)
* Drop interface key workaround for CoinRules

This was from a time when I was still a bit more optimistic that we
could just switch the main model to an upgradeable one immediately. I
think at this point it is just confusing and it makes more sense to
introduce this stuff once we actually make the full model upgradeable.