Bump SDK to latest snapshot (#59)
Seems sensible to startup the team with whatever is current now. Given
that we want to consume features fairly quickly and fast feedback, I’d
generally suggest we follow snapshots quickly unless there is a
specific reason that stops us from doing so.

The SDK upgrade script requires NPM atm so I’ve added that to the nix
env as well.