Add Admin endpoint PoC (without testing yet) (#111)
Progresses #87
This PR contains just a small checkpoint where I added a dummy Admin API command as a PoC to the class that will eventually be the Validator "Instance" class on the console side. This code still needs a lot of changes but I separated this out into a PR to not end up with a huge PR at the end.

It is not possible to test this Admin API endpoint in an integration test yet.