Add test with coins and add coin protobuf representation (#240)
Fixes #212 and progresses #206 

For authentication in the wallet tests, I added a Daml user to the wallet config but this is temporary (until we have a proper IAM wallet integration). 
In terms of code added, the majority of code in this PR (by code lines) is testing-util-code copied from Canton.