Implement tap choice and console command (#275)
This PR fixes #141 by implementing `tap` on the Daml side and offering it as a console command. 
It also standardizes the `initialize` function and adds an integration which essentially tests the basic runbook scenario we want to have, where we tap the wallet, and the SVC app and validator apps run against two different participants. 

It also improves our handling of Daml `Numeric`s