Introduce shared Contract type (#333)
* Introduce shared Contract type

This replaces all the custom Scala representations of various
contracts with something generic that exposes a subset of what the
codegen exposes. I went for a separate type to give us some option
over what and how we serialize it later (e.g., switch to protobuf
structs or whatever).

This deletes a bunch of code. Code is bad. Therefore this PR is good.

fixes #317