Switch to shared remote participant reference (#355)
Another direction towards reducing the per app boilerplate. The only
reason we cannot use RemoteParticipantReference directly is that we
need a different way to get the config. There is no reason to look
that up by name if we can just pass it along so this PR does just that
and then uses that shared class.