Make daml.yaml files compatible with SBT & Daml Studio (#374)
* Make daml.yaml files compatible with SBT & Daml Studio

Note that there is still the issue that our SBT build relies on the
files without the version name in there. This means just `daml build`
doesn’t quite work for downstream dependencies and you need to specify
the output path. Easy enough to fix but I'm tired and supposed to be off.