Add PoC of Splitwise Daml model (#530)
* Add PoC of Splitwise Daml model

This implements a first daml model for splitwise.

There is a bunch of stuff in there that I expect us to iterate on but
at least wrt to the invites & payment flow it should be reasonably
close.

It is probably easiest to start with the Daml Script in
CN.Splitwise. I’ve tried to introduce enough helpers to make the flow
easy to understand.

I’ve also put it under the experiments directory for now.