Add splitwise frontend (#656)
Note that this only implements the self-hosted mode. I'll create an
issue detailing the steps required to make provider-hosted mode work.

There also appears to be a (harmless) bug where invites keep showing up which I
think is actually a bug in the Scala code. Doesn’t really prevent this
from being usable so I'd like to merge as is.

fixes #637