Route splitwise rights directly to ledger API (#690)
This matches our third party app architecture where we want the web
frontend to interact with its own participant directly.

This is an intermediate step in #661 so I’m deliberately leaving some
stuff hacky since I expect it to change significantly. I've added
todos for all of that.

I've also completely broken the web frontend with this. I'll fix that
once the backend is fully in place to avoid too many painful iteration
cycles.