Ensure all reads in splitwise backend can be done through provider (#698)
To make our third-party app architecture tick, we need the backend to
only read as the provider.

The one case where that was an issue was reading accepted app
payments. While we could find a solution for that and support
extra observers/signatories on those, I'm deliberately not doing that
to force us to think about how we want to handle read delegation and
not just write delegation. Slightly confusingly this was actually not
used at all by our Scala tests but it is used by the frontend which
will read those contracts from its own participant.