Replace directory user apps by remote directory provider ref (#725)
This kills a whole bunch of boilerplate but more importantly it
changes the architecture to match the one used by splitwise and
generall your third party app arch where ledger API requests are
directly sent to the legder API without going through another grpc
service.

There is one caveat here: At the moment, we don’t have remote refs
that allow us to call the provider-only calls on the directory
service (e.g., collecting payment). We didn’t have that before either
so this is not a regression. In the short term, I expect we'll just
replace all of those by automation. Long-term, following our IAM
design those should be exposed on a separate grpc service anyway and
we can then get a remote ref to that.

I'd also like to kill `provider` from all the names but leaving that
out of this PR to avoid making the diff unnecessarily hard to read.