Nix-ify canton (#724)
* install canton via nix

* uncomment daml2ts

* fix builder

* typo

* almost forgot about the CI fella

* Avoid custom nix builder (#726)

Generally best practice is to only use a custom nix builder if you
need to (which is pretty much never).

* adjust readme

* Fix group join test flake (#722)

* Fix group join test flake

* use provider party

* use splitwise participant

* Add a definition for a staging cluster within scratchnet. (#727)

* refuse to run start-canton if canton is already running (#731)

* Add support for a CI job that pushes to staging on each commit to main. (#732)

* Add a flag to requests 'cncluster apply' rebuild the image prior to applying them. (#734)

* Add remote splitwise nodes (#685)

part of #661

I've kept the remote name as the simple one as that is the one that
matches our web frontend, i.e., the one we really care about. The
backend is just for admin tasks & maybe init (which we should also get
rid of)

* Reassign and update all TODO(Arne) (#737)

progresses #664 

I went through all my TODOs in the code, and either deleted them, assigned them to #736 or assigned them to TODO(tech-debt) (which I introduced because I find it easier to remember than M1-92). 

I don't think any of these TODOs are particularly urgent.

* Route splitwise rights directly to ledger API (#690)

This matches our third party app architecture where we want the web
frontend to interact with its own participant directly.

This is an intermediate step in #661 so I’m deliberately leaving some
stuff hacky since I expect it to change significantly. I've added
todos for all of that.

I've also completely broken the web frontend with this. I'll fix that
once the backend is fully in place to avoid too many painful iteration
cycles.

* Ensure all reads in splitwise backend can be done through provider (#698)

To make our third-party app architecture tick, we need the backend to
only read as the provider.

The one case where that was an issue was reading accepted app
payments. While we could find a solution for that and support
extra observers/signatories on those, I'm deliberately not doing that
to force us to think about how we want to handle read delegation and
not just write delegation. Slightly confusingly this was actually not
used at all by our Scala tests but it is used by the frontend which
will read those contracts from its own participant.

* Support provider-hosted read requests (#684)

This PR finally gets us to the point where a splitwise reference can
do reads through the provider’s backend hosted on their participant
while writes still go through their own participant.

So the splitwise reference now follows the same architecture we want
for the web frontend.

Self-hosted mode still works. In that case, you just host your backend
locally and use your own party as the provider.

Next steps: A bit of cleanup and then ~lots of sleep~ figure out how
to fix the web frontend which at this point is completely and utterly
broken.

part of #661

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>
Co-authored-by: Mike Schaeffer <40773844+mschaef-da@users.noreply.github.com>
Co-authored-by: Itai Segall <itai.segall@digitalasset.com>
Co-authored-by: Arne Gebert <arne.gebert@digitalasset.com>