Create ValidatorRights for users (#843)
* Create ValidatorRights for users

This PR creates the proper validator rights for all users and changes
the test to collect validator rewards for the validator user instead
of faking users being their own validators.

There is one slightly annoying point about making sure that the
validator can use the wallet: We need to create the WalletInstall
contract after both the wallet and the validator have been
initialized. For now, I solved this by adding a separate command. We
could automate that later.

fixes #723 fixes #492