Fix docker config in envrc (#893)
If the file exists but the key is not there jq gives us `null` not an
empty string. Somewhat confusingly jq does seem to output an empty
string when you just invoke it to a terminal so it seems to change its
behavior depending on whether it writes to a terminal or not.

I did confirm that without this I can reproduce the issue and with
this it works reliably. The --quiet is needed to drop the confirmation
prompt which doesn’t work in direnv.

fixes #891