merge envoy for wallet and splitwise frontends (#890)
The wallet and splitwise UIs each deploy their own instance of envoy proxy. This PR merges these into a single instance.

Note that there is a separate instance of envoy which we deploy within our clusters. We may consider in the future merging those as well, or at least reusing artifacts e.g. the config file.