Expose validator party from RPC endpoint and use it in wallet (#906)
fixes #878

This allows us to kill the initialize calls in both the wallet and in
the validator and generally provides a cleaner model.

There is one issue here around dependencies:

The validator app depends (deliberately) on the wallet app for access
to the codegen’d code.
With this change the wallte app now also depends on the validator app
for the connection.

Those dependencies are fine, they only depend on the API of the other
package. In our current build setup that doesn’t work nicely
though. I’ve broken that dependency cycle by keeping the connection
abstract. Eventually, we need to rethink (I suspect we may just want
to split out APIs from the implementation) but for now that seems
fine.