Ditch envoy docker support (#943)
envoy is not supported in nix for MacOS, which is why we tried using it in docker until now. That, however, introduced several issues in linux (things like iptables rules blocking its traffic, and support in CircleCI). Rather than supporting both docker and local installation, we decided to require global installation of envoy in MacOS.