Move Wallet initialize to WalletApp (#959)
* Move Wallet initialize to WalletApp

Part of #918

This has a few advantages:

1. It removes the deadlocks arising from node ordering when running
multiple nodes in one process. E.g., currently we need to be careful
to start the validator before the wallet since the wallet queries for
the validator party in its bootstrap and we only start the next node
once `initialize` has finished. Note that it does not remove the
requirement that apps need to wait for their dependencies to be up,
that’s unchanged.
2. At least in theory it speeds up tests since startup can happen in
parallel. I haven to benchmarked any of that though so no idea if it’s
significant.
3. It brings test setup closer to our cluster deployment: While we can
control initialization order in tests, we cannot in our cluster. Doing
init in parallel for all nodes therefore matches production much
better.
4. The resources created by WalletApp are now also closed by wallet
app so resource management is easier to understand.

I see two disadvantages:

1. Doing the sys.exit in CoinNode feels even shadier than doing it in
WalletApp. But that needs to go away either way and it doesn’t really
matter where it is so I think this is fine for now.
2. It arguably
makes the need for health checks even more pressing. That seems like a
good thing though, we really should just add them.

* Override timeouts in CoinNode

* doc

* Update apps/common/src/main/scala/com/daml/network/environment/CoinNode.scala

Co-authored-by: Arne Gebert <arne.gebert@digitalasset.com>

* fmt

Co-authored-by: Arne Gebert <arne.gebert@digitalasset.com>