Add less strict startup probe to domain (#972)
We've seen a few cases where the checks failed with connection
refused. My best guess is that the domain can sometimes just be a bit
slow to startup. If we kill it due to that we just keep on starting it
and it dies again until eventually it happens to be fast enough (which
so far always happened eventually) but defining a more relaxed startup
probe seems more sensible.