Deploy validator1 wallet web UI (#987)
This PR adds the wallet web UI to our cluster. It also adds a small hack to the wallet web UI where we only use auth0 when running on a secure origin. When not running on a secure origin, the "log in with auth0" button only produces a console error but otherwise does nothing.

* Deploy validator1 wallet UI

* Work around missing TLS

* Fix backend URL