Make all bootstrap initialize calls asynchronous (#985)
fixes #918

see #959 for detailed motivation and the first step that does this for
the wallet.

There are a few things that make this slightly more complicated:

1. We want to start stuff in parallel but we need to block on all apps
having initialized before we run the actual tests. I've solved that
via the healthcheck endpoint. However, that was completely broken
before so I’ve wired that up somewhat sensible. We definitely need to
do more tweaks on that but at least it works sort of sensibly now
instead of always returning false which it did before.
2. We now hit retries in the AppConnections eg for establishing
validator connections. The retries should really be info statements
instead of warn so I've tweaked the infra around that to support that.