disable dirty flag (#1016)
I give up on getting npm & damlc to behave consistently on my machine and on circleCI and produce reliable package-lock files. Since this entire issue is supposed to be resolved soon, when we bump canton version, I propose we simply disable the dirty flag for now. This should, to the best of my understanding, bring CI and the preflights back to a working state.