Fix logback appender for --log-file-appender=off (#1032)
We deliberately disable that since we collect the logs via
k8s/gcloud. However, it turns out NullAppender does not exist at
all, see
https://stackoverflow.com/questions/17610574/does-logback-have-a-nullappender. As
a result of this you get a nasty log message at the beginning of your
logs which breaks lnav and you have to split the log file.

I'll upstream the fix to Canton ofc.

While I'm at it I've also changed our dockerfiles to all use the same
logging options. Afaik the inconsistency was purely accidental. Would
be nice to factor this out eventually but given that we create these
dockerfiles by copy paste I'm not even going to bother with a tech
debt issue for that atm.