Stop VPN before uploading caches (#1038)
Our cache uploads seem to pretty much all time out when the VPN is
enabled. There is no reason to enable it for that, so this PR changes
the job to kill the VPN before running the cleanup at the end.

I tested that the job passes that way for the manual preflight check
and that it also passes if I don’t enable the VPN in the first place
while otherwise I get reproducible timeouts so optimistic this will
actually work.