Fix authentication on deployment job (#1062)
We lost devnet auth when we changed the slack notifications. You can’t
use a logic statement in parameters
https://discuss.circleci.com/t/is-it-possible-to-use-logic-statements-when-setting-parameters-in-jobs/41862
and the work arounds seem shitty so I've just enabled both for now. I
think we can live with slightly extended privileges here.

I've also turned the failure in `cncluster activate` into a failure
and removed a redirect to /dev/null. That should result in the auth
failure actually showing up in the logs instead of a later step
failing.