Introduce retry infra and use it to retry svc party alloc (#1054)
* Introduce retry infra and use it to retry svc party alloc

part of #1047 and #967

I wanted to fix our SVC app crashing as part of #967 and thought I
might as well pull in the proper infrastructure and get any possible
conflicts on this out of the way.

I changed the LocalRunBookIntegrationTest to start the SVC app before
connecting to the domain which fails without this change so retries
work properly there.

I ended up moving it to a separate CoinRetries trait since I expect we
may want to use this outside of nodes, e.g., make the automation an
instance of that. Also just seemed cleaner to not pull too much crap
into that trait.

I did not yet replace all other retries consistently. In particular,
retries on command submissions and retries for stuff like
getValidatorPartyId are not using this yet.