Drop builtin retries in command submissions (#1068)
* Switch retries in AppConnection to CoinRetries

Kills the internal retries in AppConnection and moves them to the call
site based on CoinRetries.

There is one slightly hacky part in there, see #1066.

I also bumped the retries since in our cluster deployment some stuff
still timed out.

part of #1047

* Drop builtin retries in command submissions

fixes #1047

This removes the builtin retries in command submissions.

For user-initiated requests we just fail for now. For automation we
retry. I've added that only to the SVC automation because that is not
yet switched to the new automation framework. The directory automation
is taken care off by #1049 so anything I do here just conflicts.

Note that I kept `ignoreDuplicateKeyErrors`. That really needs to be
replaced by logic that queries the contracts from stores but those
stores are not hooked up yet. It also doesn’t actually do any retries
so it’s really part of #885

.