Retry command submissions in validator init (#1087)
We saw some issues in our cluster where those failed because the SVC
participant had not yet vetted the corresponding domain. Adding a
bunch of retries should help with that. This is fairly coarse
grained. Some of those commands don't involve the SVC participant and
should be fine without this but adding it to everything in validator
init seems easier to get consistent.