Fix initialization order in CoinNode (#1164)
This is best explained with
https://gist.github.com/cocreature/e225b7000fc4fe502e68de3fb58260ea
and
https://docs.scala-lang.org/tutorials/FAQ/initialization-order.html.

In short: Overriding a definition does not affect the value in the
constructor of the superclass for certain combinations of
definitions. In our case, we access those valuse in a future started
from the constructor so this is not just broken it is
non-deterministically broken (my favorite kind of broken).

Based on the gist I included lazy vals seem like the only safe option
because you can only override it with another lazy val. Everything
else is broken in some way. You can’t have abstract lazy vals though
so some stuff is still a `def`.

I don’t think I will ever sleep soundly knowing this but hopefully
this at least fixes the issue for CoinNode.