Truncate long party IDs in UI (#1265)
When party IDs get resolved to a directory entry, we show also the party ID next to the resolved name. But that starts becoming a long string that messes up with the UI. This PR truncates this into XXXX...XXXX. The full ID can be seen in a tooltip when hovering over the shortened one.