Add activeness lookup checks for all coin operations (#1253)
This PR adds activeness lookups for all coin operations that ensure that all contracts, that need to be active for a given coin operation to succeed, are active at the time of submission. Commands that fail this check aren't sent to the ledger and instead an error is returned to the console command client.