use DirectoryEntry component in splitwise ui (#1245)
Uses the DirectoryEntry component to lookup entries and display them, similarly to the way the wallet uses that.

Note that we can not yet retire the directoryEntries class that polls for all entries, since we use that list to broadcast invites, and for auto-completion of entry names. In the future, we should add an API to the directory app that supports that capability.

Also note that browsers tend to take liberty to break strings and add newlines, so this commit adds also a matcher that compares two sentences while considering all whitespaces as equal.