Make signatories control payment collection (#1354)
Note that this choice is obsolete from a delegation perspective
as the calling code has sufficient authority to implement its body
directly. However, this choice does simplify code auditing, which
we expect to be helpful further down the line when we want to
display what rights a particular app requires.