Drop key from CoinRules (#1370)
as part of that also dropped the key from IssuingRound which I forgot
in the previous PR (I had removed all uses of that key, just forgot
the key annotation) and one stale BurnOracle template.

With that the core CC model is now completely free of keys. I'll open
separate issues for the other models to remove keys.