Add automation to collect payments to splitwise backend (#1381)
* Add automation to collect payments to splitwise backend

I've completely killed the UI for displaying payment requests with
that. We could still display them while they’re in flight but that
just doesn’t seem useful and it was completely untested. We can always
add it back later if we want to.

Note that this changes the non-multi transfer requests to also lock to
the provider. Otherwise the provider cannot collect the
contracts (without divulgence).

At this point self-hosting splitwise makes increasingly less sense and
keeping the test working is just super confusing so I've killed
that (we already killed it from the Scala model earlier).

fixes #1341