Drop keys from splitwise (#1474)
* Drop contract keys from splitwise

fixes #1373

fixes #1282

This one is slightly more complex than the other ones: We need to
enforce uniqueness for the pair (provider, user, groupId). To support
that, I switched group creation to the usual request/accept flow and
handle the dedup provider side since trusting the provider to enforce
that properly seems much more sensible than trying to enforce that in
the UI or somewhere like that.