Switch DirectoryReference to java codegen (#1487)
This switches over the CLI frontend & tests to the Java codegen.

That is the last reference for the directory files so this PR also
disables the Scala codegen for the directory app.

The extended Canton APIs are not super clean but good enough for now
imho. I expect we clean them up once we upstream our changes there.