Make subscriptions multi-domain compatible (#1518)
fixes #1366

Note that this means the renewed contract has to be moved to the public
domain for renewal. I think that’s somewhat unavoidable unless we're
willing to give up on atomic settlement unfortunately. For directory
entries that’s obviously not an issue since they're on the public
domain anyway but for other apps this seems somewhat unfortunate
(although not that big of a problem either).