Fix flakiness in frontend tests (#1550)
What seems to be happening sometimes is that the test shuts down which
first shuts down the backend and then the frontend. That can result in
an error being thrown in the frontend which proxies as an error to our
logs and then fails CI. This is the same issue we had in the "Found
unexpected errors stuff" so applying the same fix here to close stuff
in testFinished should hopefully do the trick.