Remove OpenBusiness modules (#1597)
I've completely removed Openbusiness.Expiry. We only use it for Coins
and LockedCoins and there inlining the code seems just as simple. I
expect we might end up moving those choices to CoinRules as part of
the upgrading work but for now just want to restore what is there.

We clearly still need fees but CC.Fees seems like a less confusing
module name for that.