Move wallet payment APIs to separate DAR (#1598)
Preparation for #1280. We only need to decouple the payment APIs so
having them in a separate DAR makes it a lot easier to get an
overview. Also seems like a sensible change either way to split
user-facing APIs from our internal APIs.

I've kept the payment channels as part of the internal API for now
given that we're planning to replace them by one-time payments. I
expect one-time payments will be part of the externally facing DAR.