Make task execution parallelism of automation services configurable (#1726)
Part of #790.

The expected use of configuring that parallelism is to set it to 1 in tests that require sequential exeuction, and to tune it to higher numbers for apps that require extra task handling throughput. The added infrastructure also makes it easy to pass extra configuration parameters to the automation services in case discover them.