Support disabling LockedCoin_Unlock (#1764)
Part of #1280

This only tackles unlock so we can agree on the approach using one
choice and then extend it to everything.

Note that the choice is only coupled to the interface of CoinRules (so #1760 was not
necessary for this but I think it’s good to keep it independently).
If we reference the CoinRules template instead, we would need to keep
around multiple CoinRules contracts to support Unlock of old Coin
contracts. For unlock that may be acceptable if we view locking as a
short term thing so we don’t have long-living LockedCoin contracts but
as a general pattern this seems problematic to me.