Support disabling all choices in Coin model (#1788)
* Support disabling all choices in Coin model

fixes #1280

This adds boolean flags for all choices that do not have the SVC as
controller.

The one exception to that is the choices on ValidatorRights. Those
contracts currently do not involve the SVC at all so I don’t see why
the SVC should disable them. Having the validator operator handle
those seems more sensible.

* Update canton-coin-api/daml/CC/API/V1/Coin.daml

Co-authored-by: Simon Meier <meiersi-da@users.noreply.github.com>

Co-authored-by: Simon Meier <meiersi-da@users.noreply.github.com>