Fix SBT Daml dependency issues (#1794)
fixes #1766

The issue was that we defined both Test/damlBuild and
Compile/damlBuild and the dependencies inconsistently pointed at one
of them so we built it twice with no ordering between the steps. I
verified that with a bunch of extra logging.

I ran this 3 times on CI and got no failure whereas before we pretty
consistently got one across 3 runs so hopefully this actually works.