Polish unit tests for `InMemoryAcsStore` (#1798)
Fixes #1748.

Includes removing the `sbt` command alias for `test` and `testOnly`, as these did not work with CircleCI's test-splitting while having multiple projects with tests. Instead of a command alias, we now disable specific `test` and `testOnly` task keys for all canton projects in the build.

Also removes the dummy `SvcAppSpec` unit test, as we now have a proper unit test to serve as an example.