Fix flake in user onboarding by switching to onboardWalletUser (#1809)
This switches all uses of `onboardUser` except in the wallet integration
tests to `onboardWalletUser` which does the necessary synchronization to
make sure it’s been ingested in the wallet. Not strictly necessary in
all places but doing it consistently seems better.

The alternative would be to make `onboardUser` itself
synchronous. That seems nice in theory but in practice it means we now
get circular deps where the wallet calls into the validator service
and with this change also the other way around. Not inherently
problematic but at least confusing. It also seems inconsistent since
none of our other commands wait for store ingestion so blocking here
seems weird.