Make polling trigger sequential and run only for most recent timestamp (#1866)
as the code before this change was badly behaved for handlers slower than the polling interval: the trigger would
1. run multiple of them in parallel
2. always run at least one handler instance for every polling timestamp

This removes one reason for the flake in #1848. No new test was added, as I did not find a way to write a reliable test for distinguishing the faulty behavior from the desired one: skip timestamps during which the handler was running, and process at most one handler at a time.