Canton OS dependency bump (#1887)
fixes #480
fixes #1608

Note that I didn't update to the latest Canton OS code drop, because the latest SDK release doesn't use that, but a release candidate from a private (?) branch and our tooling does not currently support that and we don't urgently need anything from it. I instead updated to the second-latest Canton OS code drop.