Switch directory service to USD payments (#1895)
fixes #1641

I thought about how to test that but tbh I can’t come up with a test
that actually tests something. We test subscriptions themselves in
unit tests so at most we would test that the subscription has the
right field set. But checking that the test checks that correctly is
just as hard as checking the actual code so that doesn’t seem
meaningful. So going without a test for now.