Factor out true `PollingTrigger` from `PollingParallelTaskExecutionTrigger` (#2170)
This provides an abstract facility for implementing daemon-like code
that regularly wakes up, polls for work, and, if yes, performs that work.
Moreover, the daemon can opt-in to do an immediate follow-up check for
more work after having completed its initial piece of work.

This is e.g. useful when working off of a task queue being filled by
another process. As more tasks might be added, while taking care of
the currently available tasks.