Remove the ubiquitous use of `QueryResult` (#2173)
We now qualify methods that return `QueryResult` with
a `WithOffset` suffix, and use the unqualified name
for the method that returns only the value. We found
that this makes the code more succinct.

We recommend to going forward add the two variants
of store methods in a demand-driven fashion, as we
already do for the `getXXX` variants of `lookupXXX`
methods.

Fixes #2012.