Add domain stores to all apps (#2248)
That should give us easy access to domain ids where we need it.

There is a fair amount of duplication in the stores & ingestion here
which I'm not super happy with. I've opened #2243 for this.

The more annoying duplication is around the listConnectedDomains
endpoint. It seems tempting to factor out a shared gRPC service for
listConnectedDomains. Unfortunately, that would just make the switch
to HTTP APIs harder so I accepted this for now and we can revisit it
after that switch has been finished.

fixes #2339