Expose participant metrics in cluster (#2259)
* Expose participant & domain metrics in cluster

They're not stored yet but you can curl them for debugging at least.

part of #2252

This also fixes an issue where we were not using canton-research for
the domain and one participant. I slightly wonder if that may actually
be the cause of the threadpool issues but I don’t have a good
explanation for why it should be.