Make splitwise & wallet aware of target domains (#2320)
This removes all calls to `getUniqueDomainId` in favor of resolutions
to specific domain alias.

Also renamed the domain to `global` following guidance from product.

At the moment, there is not yet a dedicated splitwise domain so
splitwise & global point to the same.

Currently this resolves the domains multiple times. We may eventually
want to move this to startup but there is a bit of a bootstrappping
problem where the domain ingestion needs to be started for us to do
resolution but it is started together with triggers that need the
domain id. I'd like to wait until we have per domain stores to see how
we change this.

fixes #2309