Add orchestration for per-domain automation (#2327)
part of #2323

This does not yet initiate transfers. For now, we just check that the
automation gets spun up and we log that the contract is on the wrong
domain.

This isn’t hooked up particularly nicely. I'm directly querying the
ledger API. Once we have multi-domain stores we need to revisit this
but for now this should allow us to not couple the store work and the
transfer sends too closely.