Switch testnet to weekly (#2360)
We've finally switched to pg so this should be reasonable now. We're
also starting to have actual users so changing this is becoming more
important.

I'd suggest we do a manual redeploy of testnet once today. It would
have been redeployed tomorrow anyway and our in-memory testnet from
Monday will probably explode before the end of the week anyway.

Assuming noone disagrees with this, I'll do that once this PR is merged.