Fix flake due to failed joinConsortium call in LocalRunbook tests (#2396)
I've seen this fail several times on CI now which is not
great. Ideally the request failing here would just not be an issue, I
opened https://github.com/DACH-NY/the-real-canton-coin/issues/2395 for
that but this should at least unbreak CI.

I adapted the ValidatorIntegrationTests to do the explicit sync
upfront which seems easier to get right to me and is more
consistent (some tests already did that). It is technically a tiny bit
slower. But all our regular tests work like this so I don’t see why
this should not be acceptable for the ValidatorIntegrationTests.