Connect directory UI directly to directory backend (#2460)
This removes the local reverse proxy from both validator1 in our
cluster and from the runbook and instead directly connects to the
directory backend.

We don’t actually need to set CORS headers because our backend already
does that but we do need to enable TLS for validator1 to work.

I tested this manually using the runbook in both the wallet &
directory and I tested wallet, directory & splitwise on validator1.

Unfortunately, I couldn’t figure out a reasonable way to avoid having
to edit config.js. We rely fairly heavily in our artifacts not being
cluster specific atm so we cannot easily just bake it in the
artifact. I used a URL that is broken by default rather than
connecting to the wrong cluster. so if someone forgets
it, it’s easy to debug.

Sadly no tests, the best test would be extending the preflight checks
with this but that’s pretty involved and more than I can offer this
evening.

part of #2228, leaving open for the test parts.