Fix synchronization on splitwise group joins (#2491)
The comment got at the right idea but it wasn’t actually correct. The
element do indeed not disappear. They have keys in react and the
actual dom element stays stable. What does break in the `findAll` call
which can result random garbage while we're still waiting for the
previous element and then the following `filter` calls and so on fail.

We also need to add explicitly wait for the previous join to
finish. Otherwise we can submit joins too fast and because the join is
a consuming exercise badness ensues.