Test splitwise domain -> global domain operations (#2520)
* Test splitwise domain -> global domain operations

Part of #2353

This covers the first half of the multi-domain flow where you initiate
a payment request on the private domain and then transfer it to the
public domain and approve it. It does not yet cover the other
direction. For that we need #2519 and we need to change the automation
to run on the splitwise domain rather than the global domain which
probably isn't going to fly without multi-domain stores.

There is a bunch of refactoring in here and a few small fixes to

1. Only consider stake holder contracts for ACS-style filters.
2. Only submit transfer ins where we submitted the transfer out.

* Update apps/common/src/main/scala/com/daml/network/environment/CoinLedgerConnection.scala

Co-authored-by: Stephen Compall (Actually @S11001001) <stephen.compall@digitalasset.com>

* Update apps/common/src/main/scala/com/daml/network/environment/LedgerClient.scala

Co-authored-by: Stephen Compall (Actually @S11001001) <stephen.compall@digitalasset.com>

* Update apps/common/src/main/scala/com/daml/network/environment/LedgerClient.scala

Co-authored-by: Stephen Compall (Actually @S11001001) <stephen.compall@digitalasset.com>

---------

Co-authored-by: Stephen Compall (Actually @S11001001) <stephen.compall@digitalasset.com>