Make splitwise payment collection multi-domain compatible (#2534)
* Make splitwise payment collection multi-domain compatible

fixes #2519

This is the splitwise equivalent of #1359

This is slightly more interesting than the other switches: We
performed some assertions in relation to the `Group` on collection, in
particular, we checked that both members are still part of the group.

We cannot perform those checks anymore here since the group contracts
stay on splitwise’s domain.

This is fine because the check was only a safe guard. We don’t
actually support removing anyone from the group right now.

If we do add user removal, I think the right fix here is to only
actually remove people once there are no more in-flight
TransferInProgress contracts (potentially cancelling ongoing
once). But not implementing that for now.

* Update daml/splitwise/daml/CN/Splitwise.daml

Co-authored-by: Stephen Compall (Actually @S11001001) <stephen.compall@digitalasset.com>

---------

Co-authored-by: Stephen Compall (Actually @S11001001) <stephen.compall@digitalasset.com>