Apply makefile baseimage dependencies to docker-push-force (#2548)
fixes #2533

Does this work? I tested it but we're talking about makefile so
probably not.

Does Moritz deserve to rot in hell for the use of include files in
combination with make’s YOLO scoping (yes that is the technical term)?
yes.

Is there no better way of doing this? I also implemented a
version using `eval` and 3 layers of escaping. I'm sure it comes as a
surprise but it was not better.